+++ actions/browsing/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.1.1] - 2025-01-09
+
+### Changed
+
+- Dependency versions updated
+
+### Fixed
+
+- Return type for `fill_elements` action
+
 ## [1.1.0] - 2024-11-18
 
 ### Added