+++ actions/email/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.0.1] - 2025-01-09
+
+### Changed
+
+- Dependency versions updated
+
+### Fixed
+
+- Argument handling for `send_email` action
+
 ## [1.0.0] - 2024-10-08
 
 ### Changed