+++ actions/google-mail/CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.1.3] - 2025-01-09
+
+### Changed
+
+- Dependency versions updated
 
 ## [1.1.2] - 2024-10-08
 
 ### Changed
 
 - Updated core dependencies
-
 
 ## [1.1.1] - 2024-07-31